Add files via upload
Added some diagnostic messages to a few classes trying to debug a problem that turned out to be error of throwing an exception in a call back function.  Also added a feature to slide show limiting the pictures by the number of pictures a set contains, set being files with same prefix, pic_01, pic_02 pic_03 etc.  If the set contains less that the specified number of files it will be omitted from the show.